+++ src/css-rules/no-unknown-unit/index.ts
@@ -1,0 +1,3 @@
+import noUnknownUnitRule from './rule-definition.js';
+
+export default noUnknownUnitRule;